Update theming engine, add no utendo option in settings

+++ kamekverse/views.py
@@ -718,14 +718,18 @@ def settings_site(request):
     if request.method == "POST":
         profile = Profile.objects.get(user=request.user)
         if request.POST.get("theme"):
-            theme = Theme.objects.get(tid=request.POST.get("theme"))
-            profile.theme = theme
+            theme = Theme.objects.filter(tid=request.POST.get("theme"))
+            profile.theme.set(theme)
         else:
-            profile.theme = None
+            profile.theme.clear()
+        if request.POST.get("noutendo"):
+            profile.themes_urls = ["/static/offdevice/noutendo.css"]
+        else:
+            profile.themes_urls = []
         profile.save()
     requestinfo = PageStartRoutine(request)
     layout = requestinfo["layout"]
-    themes = Theme.objects.all()
+    themes = Theme.objects.filter(is_hidden=False)
     data = {"name":settings.APP_NAME,"IS_PROD":settings.IS_PROD,"ENV_ID":settings.ENV_ID, "themes":themes}
     return render(request, f"{layout}/site_settings.html", data)
 # API views (Kamekverse's custom API, the replica of Miiverse API may be in an extension just like the console UIs if i will ever get to make it)
@@ -888,12 +892,16 @@ def api_post_toggle_yeah(request, post_id):
             post = Post.objects.get(post_id=post_id)
             user = GetAPIUser(request)
             if user.is_authenticated:
-                if Post_Yeah.objects.filter(post=post, user=user).exists():
-                    Post_Yeah.objects.get(post=post, user=user).delete()
-                    resp = JsonResponse({"status": 200, "action": "delete"})
+                if not (not settings.ALLOW_SELF_YEAH and post.creator == user):
+                    if Post_Yeah.objects.filter(post=post, user=user).exists():
+                        Post_Yeah.objects.get(post=post, user=user).delete()
+                        resp = JsonResponse({"status": 200, "action": "delete"})
+                    else:
+                        Post_Yeah.objects.create(post=post, user=user)
+                        resp = JsonResponse({"status": 200, "action": "create"})
                 else:
-                    Post_Yeah.objects.create(post=post, user=user)
-                    resp = JsonResponse({"status": 200, "action": "create"})
+                    resp = JsonResponse({"status": 403})
+                    resp.status_code = 403
             else:
                 resp = JsonResponse({"status": 403})
                 resp.status_code = 403
@@ -916,12 +924,16 @@ def api_post_toggle_nah(request, post_id):
             post = Post.objects.get(post_id=post_id)
             user = GetAPIUser(request)
             if user.is_authenticated:
-                if Post_Nah.objects.filter(post=post, user=user).exists():
-                    Post_Nah.objects.get(post=post, user=user).delete()
-                    resp = JsonResponse({"status": 200, "action": "delete"})
+                if not (not settings.ALLOW_SELF_YEAH and post.creator == user):
+                    if Post_Nah.objects.filter(post=post, user=user).exists():
+                        Post_Nah.objects.get(post=post, user=user).delete()
+                        resp = JsonResponse({"status": 200, "action": "delete"})
+                    else:
+                        Post_Nah.objects.create(post=post, user=user)
+                        resp = JsonResponse({"status": 200, "action": "create"})
                 else:
-                    Post_Nah.objects.create(post=post, user=user)
-                    resp = JsonResponse({"status": 200, "action": "create"})
+                    resp = JsonResponse({"status": 403})
+                    resp.status_code = 403
             else:
                 resp = JsonResponse({"status": 403})
                 resp.status_code = 403
